Rig4j: Cache/reuse drawing/images instead of fetching from gdoc.
Caching is done at the GDocHelper live in downloadDrawing/Image.
Usage of the cache is checked in HtmlTransfer.rewriteUrl based on the *text* content of the Element node being processed, keyed with an index based on its generated position (blog header, post header, post short, post full content).
The cache only contains the generated file path, which is checked with isFile(). We do not validate that the file has changed on disk (e.g. file hash or file length), only its presence.